SW-230 Initial implementation of users verification. Prototype of the UI. The shield for making events Public was added. Edit Profile view was added. Refactoring.

+++ Server/SmartWalk.Server/Controllers/Base/EntityBaseController.cs
@@ -81,7 +81,7 @@ namespace SmartWalk.Server.Controllers.Base
             var access = _entityService.GetEntitiesAccess();
             if (access != AccessType.AllowEdit) return new HttpUnauthorizedResult();
 
-            return View(new EntityVm { Type = (int)EntityType});
+            return View(new EntityVm { Type = (byte)EntityType});
         }
 
         [CompressFilter]
