+++ Server/SmartWalk.Server/ViewModels/ContactVm.cs
@@ -6,7 +6,7 @@ namespace SmartWalk.Server.ViewModels
     public class ContactVm
     {
         public int Id { get; set; }
-        public int Type { get; set; }
+        public byte Type { get; set; }
         public string Title { get; set; }
         public string Contact { get; set; }
