+++ Server/SmartWalk.Server/ViewModels/EntityVm.cs
@@ -14,7 +14,7 @@ namespace SmartWalk.Server.ViewModels
         }
 
         public int Id { get; set; }
-        public int Type { get; set; }
+        public byte Type { get; set; }
         public string Name { get; set; }
         public string Abbreviation { get; set; }
         public string Description { get; set; }
